+++ openapi/internal/jsoac/tool.go
@@ -1,5 +1,13 @@
 package jsoac
 
+import (
+	"github.com/jsightapi/jsight-schema-core/notations/jschema"
+
+	"github.com/stretchr/testify/require"
+
+	"testing"
+)
+
 func quotedBytes(s string) []byte {
 	bb := make([]byte, 0, len(s)+2)
 
@@ -8,4 +16,20 @@ func quotedBytes(s string) []byte {
 	bb = append(bb, '"')
 
 	return bb
+}
+
+func jsightToOpenAPI(t *testing.T, jsight, openapi string) {
+	j := jschema.New("TestSchemaName", jsight)
+	err := j.Check()
+	require.NoError(t, err)
+
+	o := New(j)
+	json, err := o.MarshalJSON()
+	require.NoError(t, err)
+
+	require.JSONEq(t, openapi, string(json), "Actual: "+string(json))
+}
+
+func strRef(s string) *string {
+	return &s
 }
